Remove protocol relative urls
We used to get protocol relative from the configuration file (to get
the same protocol in api as navigator), and re-add the protocol for
plupload (which doesn't like protocol relative urls).

In production, we now *always* use https for api and navigator. The
configuration always returns http (in dev) or https (in production)
since MEDIARITHMICS/devops-scripts#540, we can remove the manual
operations.