Introduce Audience Catalog Feature
This commit contains all components that add the ability to select
audience segments in Campaign/AdGroup edition base on an audience
catalog services.
This feature will not be available/activable until Features Switch
implementation.
Based on an ongoing backend development, CatalogService contains mocked
data that will be removed later on