Add New/Edit creative from AdGroupForm
In an AdGroupForm we can now create new creative from the dropdown menu
in the Ads section.
No sidebaritems for the creativeTypePicker form.
Once the creative is created/edited -> drawer in edition mode.
In the AdGroupCardList we can now start/reset creative audits thanks to
AuditComponent reused here.
Furthermore creative forms have been factorize in order to use the
editContentLayout and formServices such as AdGroupServiceWrapper have
been moved to the formServices repository