AudienceCatalog exclusion feature fix
Exclude people button was broken. This was due to the renderProp
pattern. Indeed the button was switching showExclude boolean in the
state, trigerring a render for the AudienceCatalogFormSection component
but rendering was blocked by the AudienceCatalogProvider component whose
renderProp prop did not change hence JSX returned by this method was not
executed.

Solution: Convert AudienceCatalogProvider into an HOC so that state
modification could trigger the redendering of impacted JSX.
Now all data fetched by AudienceCatalogProvider is now provided as props
to AudienceCatalogFormSection.