Prepare the migration of partitions to builders
The end goal is to create look-alike resources. A look-alike generates a
segment, just like audience partitions. We want to unify both as an
audience segment builder (as a super type).

A builder will have a `type` property (like other polymorphic resources)
and each specialized resource will have a `<sub_resource_name>_type`.

The goal here is to prepare the audience partitions migration. To avoid
breaking everything, the first step (this commit) is to add the new
sub type, propagate the change everywhere.

Then (in future commits), re-use the `type` field to hold the
polymorphic type (`AUDIENCE_LOOK_ALIKE` or `AUDIENCE_PARTITION`) and
code the missing bits.