Plugin Layout feature
Changes added to PluginContent, PluginEditForm and PluginField Generator
in order to take into account layout.json and csv translation files. They allow to
change the way Plugin Properties of a plugin are displayed in the form, within the
navigator.

CreateEditEmailTemplatePage was modified because of its specific behavior (it does
not use the PluginContent and PluginEditForm to download the layout.json).