Fix comments about code duplication and factorization
In CreateEditEmailTemplatePage, code was factorized between fetchCreationValue
and fetchInitialValues functions.
In PluginEditForm, factorization of code in the render whether pluginLayout was
defined or not.
In PluginFieldGenerator, internationalization of messages in fields, and use of
the property type when an unhandled field type occurs.
Several modifications in catches (for errors) and isLoading in the state.