Fix bugs plugin layout
Files like DisplayCreativeForm were modified to enable the use of plugin layouts
in the case of Creatives -> display, for example.
Code was factored in PluginEditForm. A PluginSectionGenerator class was created
in order to manage one section in the form.
Moreover, the way to display <hr /> elements between sections was changed.
McsFormSection was changed to allow a string title.