Fix monitoring timeline 'See More'
(In the monitoring timeline)
If there are activities older than the 10 displayed and yet occurring on
the same day than the last one, those are excluded from the search,
because the date in the 'to' parameter was excluded.
For now, this date is formatted as a day, so we can't use a more
precise timestamp.

To fix :
'from' now includes the oldest displayed date. The count of displayed
events on this date is stored in the state, so that we can augment the
'limit' by this amount (and then the first n-th elements are removed
from the response).