MICS-94 'No events left' showing up even if there are more
Related to #785.
The 'limit' being increased sometimes, 'hasItems' was often set to false
because the response contained more than exactly 10 items.
To fix it, the array whose length is used to set hasItems is sliced to
remove the duplicates.

Tested with the user point id '12530767-dac2-4567-9eae-0eb6deb5698e' in
FNAC-Darty.