Improve Monitoring "See More"
* For a future dev : when fetching other types of activities, it could
happen that the response was not always sorted the same way when
fetching more events. As a result, there were some events that were
removed/kept falsely because of the slice. All the events between the
response and the ones in the state are now compared, to correctly remove
the duplicates.
* To display the "See More" button, we send the same request with an
incremented limit, and compare the count ; if it's different, it means
that there are more events to be fetched.