New features (intl+pag°+formatting mainly)
* Move every text inside defineMessages, for translation purposes.
* Format the field and value in Update events, according to the specs
(style & format, eg. int/float/string). Special styling for
missing/unknown values.
* New dictionary that maps, for each resource, every one of its fields
to a message and a formatting function. It also contains a resource name
and a title, both for the History drawer. This dictionary is transmitted
through the Props, all the way from the *ActionBar.
* Changed cards' key, so that it is not re-rendered every time.
* Pagination : new 'limit' parameter, 'nextTime' is correctly set and
already displayed events are sliced from the response. To know if more
events can be fetched, a second call, transparent to the user, is made
with an incremented limit.
* Misc. style edits (padding, margins, icons, ...)