Plugin Layout V2
When displaying a plugin edition page, the plugin layout is effectively fetched and displayed.
If a section doesn't have any field, it is not displayed.
A required key (boolean value) can be added in the description of fields. This key is then used
for the purpose of form validation in navigator.
When selecting a plugin, the plugin name is displayed in the list.