PluginLayout and creative->native->new fixes
When the plugin layout doesn't have a metadata field, PluginEditSelector did not work.
It was modified to take into account null values for metadata.
When clicking on Creative->native->new->close, the response was a 404 error. It was fixed (onClose
function had a parameter and was called with a non relevant value for that parameter).
On the segment page, the layout and the thumbnail are fetched and displayed for each card.
The thumbnail is also fetched in the PluginEditSelector now.