getPluginsList fix in PluginContent and GenericPluginContent
getPluginsList in PluginContent did not fetch the small icon (via assetFileService).
It was added, using the behavior of GenericPluginContent.
Therefore, availablePlugins in the state of PluginContent was modified to become a list
of LayoutablePlugins (hence it uses a field named plugin_layout instead of a pluginLayout).
Moreover, in both PluginContent and GenericPluginContent, small_icon_asset_id is checked (to exist)
before fetching the corresponding asset file.