Fix plugin layout disabled fields
In FeedCard, when editing a plugin, properties of the layout should be disabled
when the status of the feed is active. However, several components used by plugin
layouts did not use a disabled props. This props was added and then used when
editing an active feed. Moreover, the save button is removed in this case, and
a message is printed in order to tell the user to pause the plugin in order to
edit it.