MICS-113 Trigger builder
When configuring the query of a Goal, the JSONQL builder component should now be used (the same component of Segment Builder). This will happen only in new queries and JSON-OTQL queries, for OTQL and SELECTORQL the old behavior remains active.

Moreover, the options proposed by this JSONQL builder should be only the mirrored types of UserActivity or UserEvent, depending on the client's schema.

FRONT:

ObjectLikeTypeDirectiveResource and DirectiveArgumentResource have been added to RuntimeSchema.ts.

The calls getObjectTypeDirectives, getObjectTypeDirective, getObjectTypeDirectiveArguments, getObjectTypeDirectiveArgument have been added to RuntimeSchemaService.

The class JSONQLPreview now takes another props (isTrigger: boolean) switching from Trigger builder behavior to the Segment Builder behavior.

In TriggerFormSection the renderQueryField method has been modified in order to include the JSONQLPreview.

The biggest modifications have been done in JSONQLBuilderContainer. When fetching the ObjectLikeTypes now we filter only the mirrored types of UserActivity or UserEvent if we are in trigger builder mode.