MICS-376 current_version_id can be null
When a plugin is created, is hasn't a version yet. Hence
current_version_id is possibly null when fetching plugins.
This commit fix the PluginResource interface type definition.

Moreover, it is not possible to create a plugin instance (for instance a
BidOptimizer) if the plugin hasn't a version.
Consequently, we remove those plugins from the list of available plugins
when creating a new instance.

Idealy we should ask the API to filter those plugins because it can
paginated.
Thankfully, we don't and do not need for now to use paginated call
because we only display plugin list for a specified plugin_type the max
is 25 (again for now).
So filtering is done client side and first_result option has been
disabled (with the appropriate TODO comment ^^)

For the record:
core_services=# select plugin_type,count(plugin_type) from plugin group by plugin_type;
          plugin_type           | count
--------------------------------+-------
 DISPLAY_AD_RENDERER            |    25
 VIDEO_AD_EDITOR                |     1
 EMAIL_ROUTER                   |     3
 EMAIL_TEMPLATE_EDITOR          |     1
 DISPLAY_AD_EDITOR              |     2
 DISPLAY_CAMPAIGN_EDITOR        |     3
 ATTRIBUTION_PROCESSOR          |     3
 STYLE_SHEET                    |     1
 RECOMMENDER                    |     5
 AUDIENCE_SEGMENT_EXTERNAL_FEED |    18
 EMAIL_CAMPAIGN_EDITOR          |     1
 EMAIL_TEMPLATE_RENDERER        |     4
 BID_OPTIMIZATION_ENGINE        |    13
 AUDIENCE_SEGMENT_TAG_FEED      |     3
 ACTIVITY_ANALYZER              |    22
 VIDEO_AD_RENDERER              |     1
 SCENARIO_NODE_PROCESSOR        |     1
(17 rows)