MICS-537 - Add `- Clickers`/`- Exposed` label on USER_ACTIVATION Segment
The APIs used to add a suffix to USER_ACTIVATION segment names, but it's
not doing it anymore.

Instead, we decided that the booleans returned in the segment ressource
(`exposed` / `clickers`) should be used by the client apps (ex:
Navigator) to add the suffix themselves (or do something else to show
that it's a clickers or exposed segments => the responsability of the
UX is now on the client side).

So this commit is updating all part of Navigator that are displaying
USER_ACTIVATION segments name.

- Add USER_ACTIVATION segments labels on referential export
- Add USER_ACTIVATION suffix in Query Builder search box
- Add USER_ACTIVATION suffix in Campaigns segment selection drawers
- Add USER_ACTIVATION suffix in DFv2 Timeline