MICS-559 Automation builder page
First step of the new UX for Marketing Automation.
Now the menu is splitted in two: List and Builder.
Builder is under development and will be for a time,
therefore it will be hidden under a new feature: 'automations.builder'.
The old marketing automation feature has been renamed.

In the builder section a datamart selection dialog is
shown at the begining. Depending on the datamart type
we will be triggering the new scenario builder or the
old UX. As of now, the Pionus Scenario Builder is empty.