MICS-DEV enable gitlint
Correct commit messages are something we really want to have at
mediarithmics. The carrot of a proper git history isn't enough,
`gitlint` provides the stick.

The rules are not extraordinary, and follow well known rules:

- https://tbaggery.com/2008/04/19/a-note-about-git-commit-messages.html
- https://chris.beams.io/posts/git-commit/

You can also let your editor help you (
https://github.com/tpope/vim-fugitive on my side).

The main differences in the rules are:
- force a corporate email (or the github anonymized one)
- force the jira reference in the title (MICS-1234 for example)
- forbid these freaking WIP commits

To enable the checks on the CI, use:

    gitlint --commits \
    origin/${ghprbTargetBranch:-master}..${ghprbActualCommit:-HEAD}