MICS-593 Fix Infinite Loop
The authentication method is defined in an infinite loop since its
redux-saga process. Therefore the authorizeLoop method was called on and
on which dispatched `CONNECTED_USER` events constantly. By adding a
`take()` at the end of the method, the process is stopped until a redux
event is dispatched, allowing us to perform dynamic token refreshing
without constantly checking authentication in a loop.