MICS-677 Adgroup shared audience selector bugfix
Root Cause Analysis : When using SegmentTableSelector, we don't
know whether the segment is owned or shared, so we can't compute
selectedSegmentIds correctly and it cause an error calling the fetchData
prop.

Moreover, it cannot work anyway since we can't fetch an audience service
item from its segmentId breaking the preselecting of checkboxes and the
save function of the SegmentTableSelector.

For now we fix the bug by disabling the preselecting checkboxes feature.
It means that the user cannot remove segment by unchecking it in the
SegmentTableSelector.

He can remove a segment by clicking on the icon dedicated to this