MICS-605 MICS-685 Add and delete operations
We can add a node to the automation data and the tree is redrawn
afterwards. The add node method takes a parent node and a child one.

When drag and drop is done, this information should come from
AvailableNode(dropping node) and DropNode(placeholder node).

Delete operation has been added as well.