MICS-708 Icons and drop zone
Ant icons are now supported. However this is a temporal behavior and
these icons should be included in McsIcon. Once this is done all the
AntIcon tags should be deleted.

Drop zone has been included. When approaching a drop zone, the node will
change color informing that the new node can be dropped