MICS-538 Navigator shows the wrong from for cross datamarts
RCA: Navigator wasn't doing a difference if the datamart was a Cross
Datamart or a normal one. Therefore in the settings menu you could add a
token and event rules to a cross datamart. This was not the expected
behavior.

Now we check if the datamart is a cross datamart, if it is we can only
change its name.