MICS-537 - Fix remarks
In order to centralize the code to properly format the segment names
(as there is a suffix for now for USER_ACTIVATION segments), we
introduce a new component `SegmentNameDisplay` that is now responsible
to render it.

This component can take either:
- A segmentId
- A segment resource

by props, and then it display the segment name.

Please note that this new component is not used in SegmentActionBar that
is managing the excel export of segments names => using a react
component to handle the name formatting is not possible there.

Some little tweaks were also done alongside:
- Replaced some `prop: fake_type | null` by `prop?: fake_type` to stop
using null in TS (the recommended way to go is to use undefined instead
of null in TS)
- The previous proposal to do the segment name formating everywhere
it was displayed was reverted instead of the component approach.