MICS-885 Fix resource history timeline
* New way of getting the name of a linked resource (eg. a segment).
Instead of extracting its id from the selection fetched from PostgreSQL,
I do it by parsing the history of the selection. (With the previous
behavior, we couldn't display the name when the link, aka the selection,
was deleted).
* The State of HistoryEventCard now has a map string -> string to store
the names of resources (previously it was a string, only the 1st event
was displayed with the correct name).