MICS-1054 Fix Displ. Campaign stats export
We encountered 2 issues with the stats export for a campaign of a
customer:
a. Many goals that start with the same string were attached to the
campaign
b. Some goals and/or attribution models names were containing special
chars.

When we're building the export spreadsheet, we creating sheets with:
- The goal name
- The attribution name

The issue is that excel sheet names are:
- Limited in length (31 chars)
- Can't contain special chars

So, this commit is:
- sanitizing the sheet names by removing special chars
- adding a unique id in the sheet name to avoid the situation in which
2 truncated goal name are the same and result in 2 sheets having the
same name
- Ellipsizing the attr. model if it's too long -> otherwise, the goalId
and goalName can't be displayed in the sheet name