MICS-599 Fix API call to list overlap job execution
In the back-end, the request used to retrieve overlap is based on the
external model name (PUBLIC_AUDIENCE_SEGMENT) and id (the segment id).
But several jobs can have the same external model, that's why we
retrieve USER_SEGMENT_IMPORT jobs, for instance, when getting an
overlap result.

To fix this : the API already allows to pass a job type parameter, so I
just add this to the query string parameters, to ensure that we
retrieve only first party overlaps.