MICS-1129 Update campaigns' history
* Display the create link event between campaign and campaign editor
* The resource types "EMAIL_CAMPAIGN" and "DISPLAY_CAMPAIGN" are
deprecated : all the new events are published under the resource type
"CAMPAIGN". They are still kept in resourceLinkHelpers though, so that
old events can still be displayed.
* Remove the history button for creatives : the only events we have are
the links to ad groups. But to get the ad group we need both the
campaign and the ad group ids. Those are fetched from the
RouteComponents, but all the ad groups of the history are not
necessarily included in the current campaign, so it leads to a 404
error, and the ad group appears as "(deleted)" in the history (even
though it's not).