MICS-1415 Fix export download
In the segment dashboard view we want to export the audience segment
metric stats. However the `ExportService.js` wasn't updated accordingly
to the new behaviour. Since we need the audience segment metrics in
`AudienceSegmentDashboard` and in `AudienceSegmentActionBar`, I moved
the audience segment metric retrieval from `AudienceSegmentDashboard` to
`AudienceSegmentPage` so that I can pass the datamarts as props
afterwards.