MICS-1129 Add missing resourceNames in resource history
The call to get the resource name of a linked resource is done in
componentDidMount. But sometimes, because of the pagination, the link
event is displayed in a card that already exists, so only
componentDidUpdate is called, and the linked resource's name is never
fetched (there's an empty string instead).

(Also, when an ad group or an email blast's name is null, or
empty, replace it with the id, so that we can still click on its link.)