MICS-1631 Fix banner creative
When a User is clicking on an AdRenderer box, we should create the
instance with the last plugin version.

The code was doing the hypothesis that the last plugin version is the
last one of the list returned by Core API.

This hypothesis is wrong -> this PR is checking the `current_plugin_id`
on the Plugin directly to use the proper plugin version when creating
the instance.