MICS-DEMO Fix duplicate key update
When a second variable with an existing key was added to the form, it
was assigned the ml algorithm variable id of the existing ml algorithm
id. Therefore it performed a chained update leading to an unexpected
update of the key value instead of facing the back-end bad request
(duplicate key).