MICS-2198 Improve PlacementList performance and limit CSV size
The list displayed on the placement list edit page is not using
the paginated API. Instead of calling the API when the user switch
the pagination, we’re doing an API call at the page loading and
we’re asking for all the placement_descriptors.
In MICS-2142: Import - wrong import statusClosed, I added a max_results
parameters to the query sent to the backend in order to get all
the placement_descriptors related to a placement_list. The problem
here is that, for a large number of placement_descriptors (here we
got 15000 results), the Document Object Model crash as the number
of element to handle is too large.

The best fix would be to refactor completely the form and really
use the paginated API. We would then manipulate a small number or
elements. As it is a form, we should keep in memory the changes the
 user is doing, and apply them when he saves.

The quick fix that is implemented to solve the problem is to
limit the number of line we accept for the CSV.

This commit also improve the performance when a user import a CSV
as we do not display the content of the CSV anymore (we directly
send the CSV to the backend).