MICS-1700 Implement If Node
This commit add the If Node UI to the Automation Builder.
It also fix some bugs related to the builder and implement a first
version of label on the graph.

Currently, the way the label are handled is not the best way.
We should try to use a factory/widget/model like the others
components. I tried to use this pattern for the label but it
didn't work.