MICS-FIX Omit technical name when creating a scenario
When the user create a display_campaign or send_emain node,
a default value for the technical_id field is set to "".
The user can't update this value, and it is then stored in the database.
As we check if a campaign already exist for a given technical_id,
when the user tries to save a new campaign, it throws an error.

This commit omit the technical_id when saving the scenario.
The technical_id stored is "null", and we consider that we can have
campaigns without technical_id.