MICS-2616 Display 0 and false property values
When the property of an event is 0 or false, we should display it
correctly instead of displaying the string `empty`.

The use of `!` operator in JS was preventing that.

Also, the `Tooltip` component is not working well with boolean values,
we have to convert them to String before calling them.