MICS-2691 Cant display user timeline
If no UserPointId is linked to a vectortId yet (getIdentifiers
service returns an error), `<Monitoring>` does not even try
to fetch Timeline data.

So we need to take this special case into account and try to fetch
Timeline information with only a vectorId.

The whole Timeline data well displayed only if the "last seen"
parameter is not null.

I also fixed a "0" being displayed when no segment available in
`<SegmentCard>`. And a message in `<ProfileCard>` when no profile
is available.