MICS-831 Store auth token in LocalStorage
This commit change default storage for auth tokens.
It now default to LocalStorage for security purpose

It works flawlessly for our react app.

We still have SelectorQL and Legacy Marketing Automation that are based
on angular views that use Restangular for API call so I updated
angular's AuthService to retrieve token from LocalStorage as well so
that it doesn't break those 2 views when hitting browser refresh when
user is still logged