MICS-3193 Improve consents in monitoring
The way consents are displayed in the monitoring page is improved.
A CustomObjectRenderer is created to display arbitrary objects
in navigator. Moreover, custom rendering functions can be supplied
to the CustomObjectRenderer in order to render predefined variables
of the object. Those predefined variables are characterized by the
complete variable name in the hierarchy (of the object, each layer
separated by a dot). Those variable names can be absolute or
relative.

The consents in monitoring use this CustomObjectRenderer.

Moreover, a message is added in UserChoicesCard in order to deal
with a case of empty user choices.