MICS-3204 Allow preset deletion inside feed creation form
Now that we can create presets, we should also be able to delete them.
To do so, preset cards in feed creation form now have a "close" icobn, which deletes the preset (after
confirmation from the user).

NOTE : This is only accessible to ORGANISATION_ADMIN/COMMUNITY_ADMIN/CUSTOMER_ADMIN ; to test it, you can set
`hasRightToDeletePreset` locally to return `true` all the time.