MICS-3097 Prevent scenario creation if queries are not defined
Currently, we can create a query and save a scenario even if the
related queries are not defined (i.e, there's no Query Language defined).
This aim of this fix is to prevent any scenario node save if the
user has not selected the type of query he wants to use (he can
still use an empty query).

As the scenario saving is not atomic on navigator (we save
each node successively, if one node creation fails, the ones
that have already been saved are kept), I added a check that
will explore all the tree and validate it before saving it.

If the check fails, it will return a Promise.reject with
the error message explaining why the validation fails.

This promise is handled by the AutomationBuilderPage and
the AutomationEditPage that will display an error message to
the user.