MICS-3494 Settings datamart fix
In Settings -> Datamart, in Sites and Mobile Application pages,
in some cases, no data was displayed in the table, depending on
the order in which the datamarts were fetched.

If the chosen datamart had no site or mobile application, then
an empty page was displayed, without the datamart selector.

This commit checks whether there are any sites or mobile
applications in the organisation before choosing a default
datamart or displaying an empty page.