MICS-3342 E2E Tests for Document Import
This PR aims to tests an E2E scenario for document import (activities
and profiles).

Three tests are written :

1) Successful import of profiles
2) Successful import of activities
3) Unsuccessful import of profiles due to wrong data structure in the
   ndjson file.

With this PR, I also introduced a new package : `cypress-file-upload`
which is very useful for uploading files generally with cypress.

I also added two functions for storing local storage between multiple
tests of a test suite which helps not having to login each time.