MICS-3698 Increase max_results when getting plugins
The original need was for visit analyzers, where we have 51 items, but the default count is 50 ; we increase
it to 1000 to display all items. This is in GenericPluginContent code, but I keep it to 1000 for all plugins
anyway, because no plugins views are paginated on the front-end, so we need it for all of them.