MICS-3898 Error 400 When no CompartmentId
When trying to create an audience segment for a CROSS_DATAMART, no
compartmentId is available in the "New Export" modal. When clicking on
"submit", a 400 error is returned : "Compartment id null not found".

Root cause : Compartments must be fetched with "with_source_datamarts"
option set to true for fetching compartementIds from source datamart.