MICS-3947 Fix Automations drawer
With the previous commit, we had an issue where the form drawer of a node could open indefinitely.
THis was due to an issue with the last_added + injectFeatures, now we make sure to clean the 'last_added' is set
to false when the drawer opens, to avoid infinite render loops.