MICS-4044 Validate schema fix
When validating a schema in DatamartObjectViewTab.tsx, the
schemaValidation resource in the state of the component, which
was returned by RuntimeSchemaService.validateRuntimeSchema did
not correspond to the resource returned by the platform via the
post request.

The type of the resource returned by validateRuntimeSchema is then
modified in accordance with the back end
(RuntimeSchemaValidationInfoResource).