MICS-3464 Automation - React to event basic
Add a new option to react to event form to have a 'basic' form.
This form is a list of options to be selected each one corresponding to an event ($ad_view, ...).

Both forms are stored until the 'Update' click ; it means that when alternating between both, we don't override
their content.
When clicking on 'Update', it saves the content of the current tab.