MICS-4557 Add test to check presence of LIVE event in timeline
We can't see a LIVE event in the timeline if the
user point related to a given activity is not
created yet.

This bug is a regression, it was previously fixed in december.
Adding a Cypress test will prevent new regression.

To test it, we make a call api to create an IN_SESSION
activity with a new identifier. We then check if the
activity is displayed on the timeline.