MICS-5536 Processings Segment Form
In the segment edit form, for a USER_QUERY segment, a User Choices
section to select processing activities should be displayed.
This section was not displayed, due to a problem linked to
features. This commit allows to display this section.

Furthermore, subtitles and titles of
ProcessingActivitiesFormSection are modified to depend on the
context.