MICS-5812 Tree Select
This commit removes the previous custom tree select in js, and
instead directly uses the tree select from ant design in a
custom component in ts.

Nevertheless, the previous behaviour with a list of selected items
above the tree select is kept.

This tree select is used in AudienceCatalogFormSection and
InventoryCatalogFormSection, which are contained in AdGroupForm.