MICS-5844 Fix feeds stats graph deletions display
Context :
On feeds that have at least one day containing only 'deletions' stats, those are not displayed.

RCA :
The data passed to the graph is created by iterating through daily upserts, and only then, we merge with the
deletions that occurred on the same day => days containing deletions but no upserts are never considered.

Fix :
We put all data in feedReports, not half of it.