MICS-6229 Custom action form
This commit updates the form used to edit a custom action node.
This form is composed of an input for the name of the custom
action, a select to choose between SCENARIO_CUSTOM_ACTION plugins,
and a sub-form to set the properties of the plugin. This sub-form
can use plugin layouts when they exist.