MICS-6281 Fix incorrect timestamp in segment builder
In the segment builder, when picking a date, we always set the timestamp to the start of the day, regardless of
the condition operator (Before/After/etc.), which led to incorrect results. For instance, when choosing 'Before
or Equal', the timestamp should be at 23:59, not 00:00, to include the full day.

To fix this, now I adjust the timestamp to the start/end of the day, according to the condition :
* before / afterOrEqual : go to the start of the day
* after / beforeOrEqual : go to the end of the day

So only the queries for after / beforeOrEqual should change (this is NOT retroactive, it will only be applied to
new / edited queries).

I also fixed / refactored some code according to sonarqube recommandations.