MICS-6785 Test scenarios
This commit adds a test button in the action bar of a scenario
view page, when this scenario is in ACTIVE or PAUSED status.

This button can only be seen with the feature
automations-test-scenario.

This test button triggers a drawer, which allows to test the user
point of the user through the scenario. The behavior to find the
user point of the user is similar to the one used in the timeline
page. A link to the user point timeline is given in the drawer.

If the user point cannot be found, a default message is shown.

For the moment, the Launch test button (in the drawer) does
nothing.