MICS-7127 Add to Segment Feeds
It was not possible to create a feed from a segment created with
the "Add to Segment" node, in the automation builder. This was
due to the fact that the segment was created with persisted set
to false.

Therefore, this commit allows to create those segments with
persisted set to true, enabling the creation of feeds on "Add to
segment" segments.