MICS-8097 Fix menu
In audience/feeds/list, when clicking on the source of a feed which
source type is AUTOMATION, the navigator menu on the left did not
close the audience submenu and open the automation one when reaching
the automation page.

This commit fixes this, calculating the open keys of the menu each
time the pathname is changed.