MICS-8286 Email already exists
Provide more precise error message when there is an error about email
already existing when creating a user.
Also propose an alternate more precise type hierachy for ApiResponse.
Properly using it makes for building errors which are out of the scope
of a bugfix, hence I leave the breaking changes as comment