MICS-8447 React to event upgrades for properties
Allow to add properties on sub object of event.
Before it was the component FieldNodeForm who was
used to generate the form for properties.
Now a new component is created EventPropertyForm specific
to automations. A tree select is used to be able to select
sub object of event.

In this PR I also fixed the boolean operator who was not used.