MICS-7569 Node users counters
This commit allows to display counters on nodes and exit conditions
in scenarios. Those counters show the number of user points going
through each node and exit condition, given the scenario analytics
stats.

For the moment, they are not linked to a stats service in
navigator (and an arbitrary number is shown instead for display
purposes).

Those counters are feature flagged ('automations-analytics').