MICS-8484 Feed Node configuration
When viewing a feed node configuration (in view mode or edit mode),
the name field was missing in the form.

This commit allows to see the name field in those modes, for an
already saved node. Moreover, in those modes, the name on the node
is now the same as the value in the name field (the preset name)
and not the plugin name.