MICS-8512 Split Node edge selections
In navigator, when saving a scenario, edge selections for split
nodes were not saved. Therefore, user points were stuck in split
nodes when going through a scenario.

This commit sets and saves those edge selections in order to allow
to choose the next node when dealing with a split node.