MICS-8041 Automation status refresh
In the AutomationListPage, updateAutomationStatus was not working
properly. For example, when updating the statuses of several
scenarios, after the 50th element, the statuses were not updated
since the scenarios were retrieved via a call which was not
paginated. Therefore, this commit fixes the behavior of
Set status to.

Moreover, this commit allows to refresh the table automatically
when the statuses are changed.

Finally, a new action is added to the dropdown on the right column.
It allows to activate or pause a scenario, depending on its current
status.