MICS-8489 Feed node segment
The segments created by Feed nodes in scenarios are created
automatically.
These segments are characterized by USER_LIST type and
SCENARIO_FEED feed_type.
Since they are created automatically, we don't want to display
them in the SegmentsCard in the Timeline.

This commit allows to hide such segments in the timeline.