MICS-7571 Entry node stats
This commit introduces a dashboard for entry nodes, in the viewer
mode of a scenario. This dashboard is opened in a new drawer.

This dashboard contains a date range picker, and a user points
line chart.

For the moment, due to a similarity in graphs, a similar dashboard
is also introduced for exit nodes.