MICS-8670 Fix FormSearchObject
It was impossible to configure a segment entry/exit node in
automations. A multipleValues.map was throwing an error because
multipleValues was an object and not an array in this case.
This error for multipleValues occured in handleChange in
FormSearchObject.

In fact, in this handleChange, within the treatment of the
mulitple selectProps mode, a singleValue was considered. In the
else treatment, a multipleValues was considered. Therefore, in
this commit, those treatments of the if else are exchanged.