MICS-9003 Custom Action Form
In the automation builder, the custom action form used to deal
with plugins, getting a list of custom action plugins.
This induced a problem because we could only get custom action
plugins from an organisation since the call to retrieve the plugins
was made with an organisation id parameter. Removing this parameter
would result in receiving the custom action plugins from all
organisations, which is not suited in that case.

Therefore, in this commit, the behavior of the custom action
form is changed to retrieve a list of instances instead of plugins.

This allows to have one custom action plugin with instances attached
to several organisations, each of those organisations only seeing
its own instances.