MICS-9186 Remove frequency dropdown in OTQL builder
When migrating to antd v4, we lost the ability to disable one of the dropdown choices, now we could only disable
the whole dropdown.

Since the 'at most' strategy isn't implemented in our backend, we completely remove the dropdown (more
"user-friendly" than a disabled dropdown), and only leave the 'at least' text.