MICS-9451 Campaign creatives fix
When a creative (display creative) was archived, it still appeared
in the allowed creatives to be selected for a campaign
(CreativeCardSelector), in the campaign form.

This was due to the fact that the call made to fetch creatives in
this context did not specify an archived value in its options, therefore
returning both archived and non archived creatives.

This commit fixes this, adding a false archived value in the
options of the call, therefore avoiding archived creatives to appear
in the form.