MICS-9717 Fix AggregationRenderer with @cardinality queries
Situation : Use a query with a @cardinality in it, in the Query tool ; run the query.

Issue : We have a "Oops, please try to reload the page or contact your support", because of this :
'TypeError: e.rootAggregations.buckets[0] is undefined'

To avoid this, we check if everything in the path is defined.