MICS-9045 Rework AudienceFeatureSelector
This commit introduce new resources and fonctions in order to add a new
audience feature pagination process in the AudienceFeatureSelector
drawer, in the standard segment builder. The old resources are still used
in the AudienceFeatureTable in the settings and will be remove in the
ticket MICS-9921.

The new pagination process will fetch all audience features folders at
the page création, and fetch only the require audience feature at
pagination or folder changes.