MICS-10238 Fix Overlap analysis display
I found that we have only 670 significant sizes in the data file, generated by OverlapJob.
So sometimes we dont have a correct size of the segment we are looking at.
But often we can extract this information from the overlap number,
when we look overlap of the segment with itself.

Signed-off-by: Igor Epimakhov <iepimakhov@mediarithmics.com>