MICS-10250 Redirect convertToOTQL new endpoint
The new endpoint intended to query translations is
POST /datamarts/{datamart_id}/query_translations
The method hitting the new endpoint is called translateQuery.

The payload for hitting the new endpoint should be a
QueryTranslationResource so I have adapted the components that
were hitting on the old endpoint
/datamarts/{datamart_id}/query_translations/to/otql
so they hit on the new endpoint using the new resource.

This ticket is the following to the ticket mics-9318(10356), and
will might be followed by mics-10309